requirements.txt
sentence-transformers
faiss-cpu
numpy
tqdm
python-dotenv
requests
PyMuPDF
python-pptx
